unify how we handle errors and config
