+++ martin/src/file_config.rs
@@ -3,12 +3,10 @@ use std::fmt::Debug;
 use std::mem;
 use std::path::{Path, PathBuf};
 
-use futures::TryFutureExt;
 use log::{info, warn};
 use serde::{Deserialize, Serialize};
 use url::Url;
 
-use crate::MartinResult;
 use crate::OptOneMany::{Many, One};
 use crate::config::{UnrecognizedValues, copy_unrecognized_config};
 use crate::file_config::FileError::{
@@ -16,6 +14,7 @@ use crate::file_config::FileError::{
 };
 use crate::source::{TileInfoSource, TileInfoSources};
 use crate::utils::{IdResolver, OptMainCache, OptOneMany};
+use crate::{MartinError, MartinResult};
 
 pub type FileResult<T> = Result<T, FileError>;
 
@@ -39,23 +38,6 @@ pub enum FileError {
 
     #[error(r"Unable to parse metadata in file {1}: {0}")]
     InvalidMetadata(String, PathBuf),
-
-    #[error(r"Unable to parse metadata in file {1}: {0}")]
-    InvalidUrlMetadata(String, Url),
-
-    #[error(r"Error occurred in processing S3 source uri: {0}")]
-    S3SourceError(String),
-
-    #[error(r"Unable to acquire connection to file: {0}")]
-    AcquireConnError(String),
-
-    #[cfg(feature = "pmtiles")]
-    #[error(r"PMTiles error {0:?} processing {1}")]
-    PmtError(pmtiles::PmtError, String),
-
-    #[cfg(feature = "cog")]
-    #[error(transparent)]
-    CogError(#[from] crate::cog::CogError),
 }
 
 pub trait ConfigExtras: Clone + Debug + Default + PartialEq + Send {
@@ -81,13 +63,13 @@ pub trait SourceConfigExtras: ConfigExtras {
         &self,
         id: String,
         path: PathBuf,
-    ) -> impl Future<Output = FileResult<TileInfoSource>> + Send;
+    ) -> impl Future<Output = MartinResult<TileInfoSource>> + Send;
 
     fn new_sources_url(
         &self,
         id: String,
         url: Url,
-    ) -> impl Future<Output = FileResult<TileInfoSource>> + Send;
+    ) -> impl Future<Output = MartinResult<TileInfoSource>> + Send;
 }
 
 #[derive(Clone, Debug, Default, PartialEq, Serialize, Deserialize)]
@@ -244,9 +226,7 @@ pub async fn resolve_files<T: SourceConfigExtras>(
     cache: OptMainCache,
     extension: &[&str],
 ) -> MartinResult<TileInfoSources> {
-    resolve_int(config, idr, cache, extension)
-        .map_err(crate::MartinError::from)
-        .await
+    resolve_int(config, idr, cache, extension).await
 }
 
 async fn resolve_int<T: SourceConfigExtras>(
@@ -254,7 +234,7 @@ async fn resolve_int<T: SourceConfigExtras>(
     idr: &IdResolver,
     cache: OptMainCache,
     extension: &[&str],
-) -> FileResult<TileInfoSources> {
+) -> MartinResult<TileInfoSources> {
     let Some(cfg) = config.extract_file_config(cache)? else {
         return Ok(TileInfoSources::default());
     };
@@ -277,7 +257,10 @@ async fn resolve_int<T: SourceConfigExtras>(
                 let can = source.abs_path()?;
                 if !can.is_file() {
                     // todo: maybe warn instead?
-                    return Err(InvalidSourceFilePath(id.to_string(), can));
+                    return Err(MartinError::FileError(InvalidSourceFilePath(
+                        id.to_string(),
+                        can,
+                    )));
                 }
 
                 let dup = !files.insert(can.clone());
@@ -320,7 +303,9 @@ async fn resolve_int<T: SourceConfigExtras>(
             } else if path.is_file() {
                 vec![path]
             } else {
-                return Err(InvalidFilePath(path.canonicalize().unwrap_or(path)));
+                return Err(MartinError::from(InvalidFilePath(
+                    path.canonicalize().unwrap_or(path),
+                )));
             };
             for path in dir_files {
                 let can = path.canonicalize().map_err(|e| IoError(e, path.clone()))?;
