+++ martin/src/mbtiles/error.rs
@@ -1,0 +1,8 @@
+#[derive(thiserror::Error, Debug)]
+pub enum MbtilesError {
+    #[error(r"Unable to acquire connection to file: {0}")]
+    AcquireConnError(String),
+
+    #[error(transparent)]
+    MbtilesLibraryError(#[from] mbtiles::MbtError),
+}
